+++ src/effects/feats/deflectMissilesAttack.js
@@ -1,0 +1,5 @@
+export function deflectMissilesAttackEffect(document) {
+  setProperty(document, "system.activation.type", "reactionmanual");
+  setProperty(document, "system.range.long", 60);
+  return document;
+}